requirements.txt
dash==2.6.1
dash_bootstrap_components==1.1.0
Flask==2.2.2
Flask_RESTful==0.3.9
mysql_connector_repackaged==0.3.1
networkx==3.2.1
numpy==1.22.3
pandas==1.4.2
plotly==5.8.0
PyGithub==2.1.1
pytz==2019.3
requests==2.22.0
scikit_learn==1.3.2
SQLAlchemy==1.2.19
urllib3==1.25.8
xmltodict==0.13.0
